Refactor NetworkingLayer into Separate Extensions
Marked each apiProvider instance to `private(set)` vs `private`
